Deleted int cast.

+++ src/codeu/chat/client/commandline/Chat.java
@@ -134,7 +134,7 @@ public final class Chat {
           System.out.println("Server Information:");
           System.out.format("  Start Time : %s\n", info.startTime.toString());
           System.out.format("  Time now   : %s\n", Time.now());
-          System.out.format("  Duration   : %s sec\n", (int) (Time.duration(info.startTime,
+          System.out.format("  Duration   : %s sec\n", (Time.duration(info.startTime,
                 Time.now()).inMs() / 1000));
           System.out.println("Version: " + info.version);
         }
